Fix packages broken by node10 deprecation
