+++ package.json
@@ -37,14 +37,14 @@
         "@definitelytyped/header-parser": "latest",
         "@definitelytyped/typescript-versions": "latest",
         "@definitelytyped/utils": "latest",
-        "danger": "^11.2.3",
-        "dprint": "^0.49.0",
+        "danger": "^11.3.1",
+        "dprint": "^0.49.1",
         "eslint-plugin-jsdoc": "^44.2.7",
         "husky": "^8.0.3",
-        "lint-staged": "^15.2.0",
-        "remark-cli": "^12.0.0",
-        "remark-gfm": "^4.0.0",
-        "remark-validate-links": "^13.0.0",
+        "lint-staged": "^15.5.2",
+        "remark-cli": "^12.0.1",
+        "remark-gfm": "^4.0.1",
+        "remark-validate-links": "^13.1.0",
         "typescript": "next"
     },
     "type": "module"
