+++ types/cypress-dotenv/package.json
@@ -1,16 +1,19 @@
 {
     "private": true,
     "name": "@types/cypress-dotenv",
-    "version": "2.0.9999",
+    "version": "3.0.9999",
     "projects": [
         "https://github.com/morficus/cypress-dotenv"
     ],
     "dependencies": {
-        "@types/node": "*",
-        "dotenv": "^8.2.0"
+        "@types/node": "*"
+    },
+    "peerDependencies": {
+        "dotenv": ">=10.x"
     },
     "devDependencies": {
-        "@types/cypress-dotenv": "workspace:."
+        "@types/cypress-dotenv": "workspace:.",
+        "dotenv": ">=13"
     },
     "owners": [
         {
