+++ types/dotenv-defaults/package.json
@@ -1,13 +1,13 @@
 {
     "private": true,
     "name": "@types/dotenv-defaults",
-    "version": "2.0.9999",
+    "version": "5.0.9999",
     "projects": [
         "https://github.com/mrsteele/dotenv-defaults#readme"
     ],
     "dependencies": {
         "@types/node": "*",
-        "dotenv": "^8.2.0"
+        "dotenv": "^14.0.0"
     },
     "devDependencies": {
         "@types/dotenv-defaults": "workspace:."
