+++ types/dotenv-safe/package.json
@@ -1,16 +1,19 @@
 {
     "private": true,
     "name": "@types/dotenv-safe",
-    "version": "8.1.9999",
+    "version": "9.1.9999",
     "projects": [
         "https://github.com/rolodato/dotenv-safe"
     ],
     "dependencies": {
-        "@types/node": "*",
-        "dotenv": "^8.2.0"
+        "@types/node": "*"
     },
     "devDependencies": {
-        "@types/dotenv-safe": "workspace:."
+        "@types/dotenv-safe": "workspace:.",
+        "dotenv": ">=13"
+    },
+    "peerDependencies": {
+        "dotenv": ">=8.2.0"
     },
     "owners": [
         {
