+++ types/expect-puppeteer/tsconfig.json
@@ -5,6 +5,7 @@
             "es6",
             "dom"
         ],
+        "esModuleInterop": true,
         "noImplicitAny": true,
         "noImplicitThis": true,
         "strictFunctionTypes": true,
