+++ types/fs-extra/tsconfig.json
@@ -1,6 +1,6 @@
 {
     "compilerOptions": {
-        "module": "commonjs",
+        "module": "node16",
         "lib": [
             "es6"
         ],
@@ -15,6 +15,6 @@
     "files": [
         "index.d.ts",
         "test/fs-extra-tests.ts",
-        "test/fs-extra-tests-module.ts"
+        "test/fs-extra-tests-module.mts"
     ]
 }
