+++ types/ltx/lib/clone.d.ts
@@ -1,3 +1,4 @@
-import clone from "../src/clone";
+import { Node } from "./Element.js";
 
+declare function clone<T extends Node>(el: T): T;
 export = clone;
