+++ types/ltx/lib/createElement.d.ts
@@ -1,3 +1,8 @@
-import createElement from "../src/createElement";
+import Element = require("./Element.js");
 
+declare function createElement(
+    name: string,
+    attrs?: string | { [attrName: string]: any },
+    ...children: Element.Node[]
+): Element;
 export = createElement;
