+++ types/ltx/lib/equal.d.ts
@@ -1,4 +1,6 @@
-import equal from "../src/equal";
+import Element = require("./Element.js");
 
-export default equal;
-export * from "../src/equal";
+export function nameEqual(a: Element, b: Element): boolean;
+export function attrsEqual(a: Element, b: Element): boolean;
+export function childrenEqual(a: Element, b: Element): boolean;
+export default function equal(a: Element, b: Element): boolean;
