+++ types/ltx/lib/escape.d.ts
@@ -1,1 +1,4 @@
-export * from "../src/escape";
+export function escapeXML(s: string): string;
+export function unescapeXML(s: string): string;
+export function escapeXMLText(s: string): string;
+export function unescapeXMLText(s: string): string;
