+++ types/ltx/lib/package.json
@@ -1,0 +1,1 @@
+{ "type": "commonjs" }
