+++ types/ltx/lib/parse.d.ts
@@ -1,3 +1,5 @@
-import parse from "../src/parse";
+import Element = require("./Element.js");
+import Parser = require("./Parser.js");
 
+declare function parse(data: string, options?: Parser.ParserOptions | typeof Parser): Element;
 export = parse;
