+++ types/ltx/lib/parsers.d.ts
@@ -1,3 +1,5 @@
-import parsers from "../src/parsers";
+import Parser = require("./Parser.js");
+
+declare const parsers: Array<typeof Parser>;
 
 export = parsers;
