+++ types/ltx/lib/parsers/libxmljs.d.ts
@@ -1,3 +1,4 @@
-import SaxLibxmljs from "../../src/parsers/libxmljs";
+import Parser = require("../Parser.js");
 
+declare class SaxLibxmljs extends Parser {}
 export = SaxLibxmljs;
