+++ types/ltx/lib/parsers/ltx.d.ts
@@ -1,3 +1,4 @@
-import SaxLtx from "../../src/parsers/ltx";
+import Parser = require("../Parser.js");
 
+declare class SaxLtx extends Parser {}
 export = SaxLtx;
