+++ types/ltx/lib/stringify.d.ts
@@ -1,3 +1,4 @@
-import stringify from "../src/stringify";
+import Element = require("./Element.js");
 
+declare function stringify(el: Element, indent?: number, level?: number): string;
 export = stringify;
