+++ types/ltx/lib/tag.d.ts
@@ -1,3 +1,6 @@
-import tag from "../src/tag";
+import Element = require("./Element.js");
+import tagString = require("./tagString.js");
 
+declare function tag(...args: Parameters<typeof tagString>): Element;
 export = tag;
+
