+++ types/ltx/src/equal.d.ts
@@ -1,6 +1,4 @@
-import Element from "./Element";
+import equal = require("../lib/equal.js");
 
-export function nameEqual(a: Element, b: Element): boolean;
-export function attrsEqual(a: Element, b: Element): boolean;
-export function childrenEqual(a: Element, b: Element): boolean;
-export default function equal(a: Element, b: Element): boolean;
+export default equal.default;
+export * from "../lib/equal.js";
