+++ types/ltx/src/escape.d.ts
@@ -1,4 +1,1 @@
-export function escapeXML(s: string): string;
-export function unescapeXML(s: string): string;
-export function escapeXMLText(s: string): string;
-export function unescapeXMLText(s: string): string;
+export * from "../lib/escape.js";
