+++ types/ltx/src/package.json
@@ -1,0 +1,1 @@
+{ "type": "module" }
