+++ types/ltx/src/parsers/sax-js.d.ts
@@ -1,3 +1,3 @@
-import Parser from "../Parser";
+import SaxSaxjs = require("../../lib/parsers/sax-js.js");
 
-export default class SaxSaxjs extends Parser {}
+export default SaxSaxjs;
