+++ types/ltx/tsconfig.json
@@ -1,6 +1,6 @@
 {
     "compilerOptions": {
-        "module": "commonjs",
+        "module": "node16",
         "lib": [
             "es6"
         ],
@@ -14,6 +14,6 @@
     },
     "files": [
         "index.d.ts",
-        "ltx-tests.ts"
+        "ltx-tests.mts"
     ]
 }
